rename deepmil by container.py
